fix: fix data type mismatch

+++ lib/models/weather_model.dart
@@ -3,9 +3,9 @@ import 'dart:convert';
 class WeatherModel {
   final double currentTemp;
   final String currentSky;
-  final double currentPressure;
+  final int currentPressure;
   final double currentWindSpeed;
-  final double currentHumidity;
+  final int currentHumidity;
   WeatherModel({
     required this.currentTemp,
     required this.currentSky,
@@ -17,9 +17,9 @@ class WeatherModel {
   WeatherModel copyWith({
     double? currentTemp,
     String? currentSky,
-    double? currentPressure,
+    int? currentPressure,
     double? currentWindSpeed,
-    double? currentHumidity,
+    int? currentHumidity,
   }) {
     return WeatherModel(
       currentTemp: currentTemp ?? this.currentTemp,
